Move actions inheritors of NetworkModification to iidm-modification
