+++ action/action-util/src/main/groovy/com/powsybl/action/util/PhaseShifterFixedTapModificationExtension.groovy
@@ -9,6 +9,7 @@ package com.powsybl.action.util
 import com.google.auto.service.AutoService
 import com.powsybl.action.dsl.spi.DslModificationExtension
 import com.powsybl.iidm.modification.NetworkModification
+import com.powsybl.iidm.modification.PhaseShifterSetAsFixedTap
 
 @AutoService(DslModificationExtension.class)
 class PhaseShifterFixedTapModificationExtension implements DslModificationExtension {
