+++ iidm/iidm-modification/pom.xml
@@ -77,6 +77,10 @@
             <artifactId>powsybl-iidm-test</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>com.powsybl</groupId>
+            <artifactId>powsybl-loadflow-api</artifactId>
+        </dependency>
     </dependencies>
 
 </project>
