+++ iidm/iidm-modification/src/main/java/com/powsybl/iidm/modification/CloseSwitch.java
@@ -4,24 +4,22 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.action.util;
+package com.powsybl.iidm.modification;
 
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.computation.ComputationManager;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.Switch;
-import com.powsybl.iidm.modification.NetworkModification;
 
 import java.util.Objects;
 
 /**
- * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
+ * @author Mathieu Bague <mathieu.bague at rte-france.com>
  */
-public class OpenSwitch implements NetworkModification {
-
+public class CloseSwitch implements NetworkModification {
     private final String switchId;
 
-    OpenSwitch(String switchId) {
+    CloseSwitch(String switchId) {
         this.switchId = Objects.requireNonNull(switchId);
     }
 
@@ -36,6 +34,6 @@ public class OpenSwitch implements NetworkModification {
         if (sw == null) {
             throw new PowsyblException("Switch '" + switchId + "' not found");
         }
-        sw.setOpen(true);
+        sw.setOpen(false);
     }
 }
