+++ iidm/iidm-modification/src/main/java/com/powsybl/iidm/modification/ConnectGenerator.java
@@ -4,21 +4,50 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.action.util;
+package com.powsybl.iidm.modification;
 
+import com.powsybl.commons.PowsyblException;
+import com.powsybl.computation.ComputationManager;
 import com.powsybl.iidm.network.Bus;
 import com.powsybl.iidm.network.Generator;
+import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.Terminal;
+
+import java.util.Objects;
 
 /**
  * @author Olivier Perrin <olivier.perrin at rte-france.com>
  */
-public final class GeneratorUtil {
+public final class ConnectGenerator implements NetworkModification {
 
-    private GeneratorUtil() {
+    private final String generatorId;
+
+    public ConnectGenerator(String generatorId) {
+        this.generatorId = generatorId;
     }
 
-    public static void connectGenerator(Generator g) {
+    @Override
+    public void apply(Network network, ComputationManager computationManager) {
+        apply(network);
+    }
+
+    @Override
+    public void apply(Network network) {
+        Generator g = network.getGenerator(generatorId);
+        if (g == null) {
+            throw new PowsyblException("Generator '" + generatorId + "' not found");
+        }
+
+        connect(g);
+    }
+
+    static void connect(Generator g) {
+        Objects.requireNonNull(g);
+
+        if (g.getTerminal().isConnected()) {
+            return;
+        }
+
         Terminal t = g.getTerminal();
         t.connect();
         if (g.isVoltageRegulatorOn()) {
