+++ iidm/iidm-modification/src/main/java/com/powsybl/iidm/modification/GeneratorModification.java
@@ -4,13 +4,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.action.util;
+package com.powsybl.iidm.modification;
 
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.computation.ComputationManager;
 import com.powsybl.iidm.network.Generator;
 import com.powsybl.iidm.network.Network;
-import com.powsybl.iidm.modification.NetworkModification;
 
 import java.util.Objects;
 
@@ -72,7 +71,7 @@ public class GeneratorModification implements NetworkModification {
     private void changeConnectionState(Generator g, boolean connect) {
         if (connect) {
             if (!g.getTerminal().isConnected()) {
-                GeneratorUtil.connectGenerator(g);
+                ConnectGenerator.connect(g);
             }
         } else {
             if (g.getTerminal().isConnected()) {
