+++ iidm/iidm-modification/src/main/java/com/powsybl/iidm/modification/PhaseShifterSetAsFixedTap.java
@@ -1,35 +1,27 @@
 /**
- * Copyright (c) 2019, RTE (http://www.rte-france.com)
+ * Copyright (c) 2017, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.action.util;
+package com.powsybl.iidm.modification;
 
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.computation.ComputationManager;
 import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.PhaseTapChanger;
 import com.powsybl.iidm.network.TwoWindingsTransformer;
-import com.powsybl.iidm.modification.NetworkModification;
 
 import java.util.Objects;
 
-/**
- * @author Hamou AMROUN <hamou.amroun at rte-france.com>
- */
-public class PhaseShifterShiftTap implements NetworkModification {
+public class PhaseShifterSetAsFixedTap implements NetworkModification {
 
     private final String phaseShifterId;
-    private final int tapDelta;
+    private final int tapPosition;
 
-    public PhaseShifterShiftTap(String phaseShifterId, int tapDelta) {
+    public PhaseShifterSetAsFixedTap(String phaseShifterId, int tapPosition) {
         this.phaseShifterId = Objects.requireNonNull(phaseShifterId);
-        this.tapDelta = tapDelta;
-    }
-
-    public int getTapDelta() {
-        return tapDelta;
+        this.tapPosition = tapPosition;
     }
 
     @Override
@@ -44,17 +36,11 @@ public class PhaseShifterShiftTap implements NetworkModification {
         if (phaseShifter == null) {
             throw new PowsyblException("Transformer '" + phaseShifterId + "' not found");
         }
-        PhaseTapChanger phaseTapChanger = phaseShifter.getPhaseTapChanger();
-        if (phaseTapChanger == null) {
+        if (!phaseShifter.hasPhaseTapChanger()) {
             throw new PowsyblException("Transformer '" + phaseShifterId + "' is not a phase shifter");
         }
-        adjustTapPosition(phaseTapChanger);
-        phaseTapChanger.setRegulating(false);
-        phaseTapChanger.setRegulationMode(PhaseTapChanger.RegulationMode.FIXED_TAP);
-    }
-
-    private void adjustTapPosition(PhaseTapChanger phaseTapChanger) {
-        phaseTapChanger.setTapPosition(Math.min(Math.max(phaseTapChanger.getTapPosition() + tapDelta,
-                phaseTapChanger.getLowTapPosition()), phaseTapChanger.getHighTapPosition()));
+        phaseShifter.getPhaseTapChanger().setTapPosition(tapPosition);
+        phaseShifter.getPhaseTapChanger().setRegulating(false);
+        phaseShifter.getPhaseTapChanger().setRegulationMode(PhaseTapChanger.RegulationMode.FIXED_TAP);
     }
 }
