+++ iidm/iidm-modification/src/test/java/com/powsybl/iidm/modification/LoadFlowBasedPhaseShifterOptimizerConfigTest.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.action.util;
+package com.powsybl.iidm.modification;
 
 import com.google.common.jimfs.Configuration;
 import com.google.common.jimfs.Jimfs;
